requirements.txt
feedparser==5.2.1
lxml==4.4.1
python-telegram-bot==11.1.0